♻️ カスタムフックを追加し、データ取得ロジックを整理した
